Model同服务端一致性的修改，并调整商家版首页展现方式

+++ LBModels/Product/ProductBase.h
@@ -12,7 +12,7 @@
 @interface ProductBase : NSObject
 
 //id
-@property (assign,nonatomic) NSInteger id;
+@property (assign,nonatomic) long id;
 
 //商品名称
 @property (retain,nonatomic) NSString *name;
